fix: extract logContextStorage into lightweight context module to fix webpack build
Extract LogContextStore, SimpleAsyncLocalStorage, logContextStorage,
runWithLogContext, and getLogContext from logging/index.ts into a new
lightweight context.ts module that has zero Node.js dependencies.

This breaks the webpack import chain:
  certificates/page.tsx → api.ts → logging/index.ts → node:async_hooks
  (fails on client bundle because node:async_hooks is server-only)

Now:
  certificates/page.tsx → api.ts → logging/context.ts (no Node.js deps)

logging/index.ts imports from context.ts and layers on Node's
AsyncLocalStorage enhancement only when available at runtime.
